CDAP-16301 handle failures better
Added better error handling. The plugin can fail in several ways:

1. DDL failure. This will throw the exception as-is and let the
app handle retries, up to the pipeline limit.

2. Offset commit failure. The plugin will retry up to the pipeline
limit, then throw a DeltaFailureException, which lets the app know
not to retry and transition to failed state.

3. Failure to write changes to GCS. The failure will be thrown
as-is, letting the pipeline rewind state to the last offset and
retry. This allows the plugin to stream changes directly to GCS
without buffering them in memory.

4. Failure to load GCS changes into staging table. The job will
be retried up to the pipeline limit, then throw a
DeltaFailureException, which lets the app know not to retry and
transition to a failed state.

5. Failure to merge changes from staging table into target table.
The job will be retried up to the pipeline limit, then throw a
DeltaFailureException.

Also added a unit test.