CDAP-16575 add an option to require manual intervention for drops
Added an option that will cause the plugin to fail and retry
whenever it sees a drop table or drop database event and the
table/dataset exists. This means dropping those resources will
require manual intervention, but may be desirable for users that
with to ensure that data is not accidentally deleted.